requirements.txt
fastapi>=0.110.0
pydantic>=2.6.0
pydantic-settings>=2.2.0
uvicorn[standard]>=0.28.0
langgraph>=0.0.30
langchain>=0.1.13
langchain-core>=0.1.33
langchain-google-genai>=1.0.1
pinecone-client>=3.1.0
streamlit>=1.32.0
promptfoo>=0.53.0
pytest>=8.1.1
pytest-asyncio>=0.23.6
sqlmodel>=0.0.16
